Very Hacky Lazy Reflow (Don't use in production!)
Just to get some ideas about how to implement it. I am not sure where
I should put calls to perform the full reflow. I guess these would be
the places where the full grid is needed, for example when browsing
the full history of the pane.

The changes here are super hacky and probably incorrect.