Dynamic resplit reconfiguration [2]
It is a common occurance that users of tmux create a two-level split such as the
following:

```
+---------+---------+
|         |         |
|    0    |    1    |
|         |         |
|         |         |
+---------+---------+
|         |         |
|    2    |    3    |
|         |         |
|         |         |
+---------+---------+
```

An hour later the user has some long output in `3` and wants to resize it
upward, also decreasing the height of `1`. However, what it would also resize
`2`, depending on whether the first split was horizontal or vertical.

This commit adds on the implementation of a mouse-based resize, so that the
splits reconfigure automatically to accomodate the situation, without any
apparent change, as if the splits were 'right' from the start.

Effectively, all attempts to resize a border directly starting from the example
above, will change only the two panes affected, giving the user identical
behavior regardless of how the splits were originally made.

There are cases in which this change in a no-op due to some particular pane
configurations, and the way the algorithm for resplitting works. However,
it should work on most cases, especially in a checkers-style setup (i.e.
a straight NxM).

Open questions:

- Should this behavior made configurable?
- Should it also apply to resizing not via the mouse?
- Is there a better implemenation for it?